requirements.txt
pydantic
langchain >= 0.1.13
langchain-community >= 0.0.20
langchain-experimental >= 0.0.50
langchain-core
langchain-groq
langchain_huggingface
python-dotenv 
PyPDF2
chromadb
faiss-cpu
boto3
watchtower
sentence-transformers
fastapi==0.114.0
uvicorn==0.30.6
python-multipart
python-json-logger
psycopg2-binary 
sqlalchemy
